server/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
gunicorn>=21.2
python-decouple>=3.8
